got patch: use ints for line offsets instead of longs
ints have the advantage that their size is more likely to be the same
across the various architecture supported by OpenBSD, thus introducing
less possible differences.

INT_MAX is still (at least) a few order of magnitudes higher than the
patches we dealt with (even abnormal ones.)

suggested by stsp@

+++ lib/got_lib_privsep.h
@@ -618,10 +618,10 @@ struct got_imsg_patch {
  * Structure for GOT_IMSG_PATCH_HUNK data.
  */
 struct got_imsg_patch_hunk {
-	long	oldfrom;
-	long	oldlines;
-	long	newfrom;
-	long	newlines;
+	int	oldfrom;
+	int	oldlines;
+	int	newfrom;
+	int	newlines;
 };
 
 struct got_remote_repo;
